Modified architectural components; Separated step_pandas and step_other example;
