+++ .gitignore
@@ -1,3 +1,9 @@
 venv/
 
 datasets/
+
+__pycache__/
+
+*.swp
+
+.DS_STORE
