+++ etl_simplesvet/hook.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+class Hook(ABC):
+
+    @abstractmethod
+    def connect(self):
+        pass
