+++ etl_simplesvet/ingester.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class Ingester:
+    @abstractmethod
+    def ingest(self):
+        pass
