+++ etl_simplesvet/persister.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+class Persister(ABC):
+
+    @abstractmethod
+    def persist(self):
+        pass
+
