+++ etl_simplesvet/pipeline.py
