+++ etl_simplesvet/step.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+class Step(ABC):
+   @abstractmethod
+   def run():
+       pass
+
