+++ etl_simplesvet/steps/step_other.py
@@ -1,0 +1,6 @@
+from etl_simplesvet.step import Step
+
+class StepOther(Step):
+    def run(self):
+        print("Other Step to run")
+
