+++ etl_simplesvet/steps/step_pandas.py
@@ -1,13 +1,10 @@
-from hook import HookPandas
-from ingester import IngesterPandas
-from persister import PersisterPandas
-from transformer import TransformerPandas
+from etl_simplesvet.hooks.hook import HookPandas
+from etl_simplesvet.ingesters.ingester import IngesterPandas
+from etl_simplesvet.persisters.persister import PersisterPandas
+from etl_simplesvet.transformers.transformer import TransformerPandas
+from etl_simplesvet.step import Step
 
-class StepOther:
-    def run(self):
-        print("Other Step to run")
-
-class StepPandas:
+class StepPandas(Step):
     def run(self):
         input_file_name = "datasets/Vendas.csv"
         output_file_name = "datasets/Vendas_out.csv"
@@ -23,3 +20,4 @@ class StepPandas:
         print(df, df.columns, len(df))
         persister.persist(df)
         return True
+
