+++ etl_simplesvet/transformer.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+class Transformer(ABC):
+
+    @abstractmethod
+    def transform(self):
+        pass
